v1.01
Added a "Load Dictionary" button and offloaded this whole process to the
main thread. This is mostly so that users get immediate feedback as to
whether their dictionary file is problematic *before* having to make
like 4 extra clicks and get the ball rolling.

Also added a ton more comments to keep some of the coding decisions a
bit more clear, as well as to help identify certain blocks of code more
easily.